requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu 

cloudpickle==2.0.0
Flask==2.0.3
numpy==1.16.6
Pillow==7.1.2
scikit-image==0.17.2
imagecodecs==2020.5.30
tifffile==2020.5.30

setuptools<45.0.0
pyinstaller==3.5

scipy==1.4.1
matplotlib<3.2

onnxruntime==1.7.0

#XXX: try to remove those
pytorch_lightning==1.0.6
scikit-learn==0.23.2
imbalanced-learn==0.6.2

